HACK: fix setup.py for recent Ubuntu releases
This is a hack to make the setup.py work sufficiently for recent Ubuntu
releases. It's a hack because gerbmerge should really be installed at
/usr/local/lib, where this will put junk in both /usr/lib and
/usr/local/lib. This seems to be an issue with the changes that happened
starting in Python 2.6 where site-packages began to be phased out. This
could be improved by ensuring no files are placed in /usr/lib and all
files are placed in /usr/local/lib. For this to work, the
gerbmerge-starting script would also need to be redirected properly.

Seeing as how this project hasn't been developed in many years, it seems
a working hack is sufficient for our purposes.

This patch has not been tested for any impact it might have on Windows
installation or installation on other distributions that aren't Ubuntu.